moved to folders and broke header and filedata into own files.
