+++ Core/Archive/Compression/LZSS.cs
