+++ Core/Core.csproj
@@ -137,7 +137,9 @@
     <Compile Include="Archive\ArchiveBase.cs" />
     <Compile Include="Archive\Archives.cs" />
     <Compile Include="Archive\ArchiveWorker.cs" />
-    <Compile Include="Archive\ArchiveZZZ.cs" />
+    <Compile Include="Archive\ZZZ\ArchiveZZZ.cs" />
+    <Compile Include="Archive\ZZZ\FileData.cs" />
+    <Compile Include="Archive\ZZZ\Header.cs" />
     <Compile Include="AV\Audio.cs" />
     <Compile Include="AV\Movie\Directories.cs" />
     <Compile Include="AV\Movie\Files.cs" />
@@ -877,7 +879,7 @@
     <Compile Include="Kernel\Kernel_bin.Weapons_Data.cs" />
     <Compile Include="Kernel\Kernel_bin.Zell_limit_break_parameters.cs" />
     <Compile Include="Loc.cs" />
-    <Compile Include="Archive\LZSS.cs" />
+    <Compile Include="Archive\Compression\LZSS.cs" />
     <Compile Include="Memory.cs" />
     <Compile Include="Menu\Images\Cards.cs" />
     <Compile Include="Menu\Images\Cards.ID.cs" />
